move `CalculatePow`
